Make errors while resolving session handled correctly
